requirements.txt
tensorflow==2.16.1
numpy==1.26.4
matplotlib==3.8.4
Pillow==10.4.0
streamlit==1.38.0
protobuf<5
